Reorganize contributions into a tree structure. Add mobile layout.
