+++ src/app/management/pages/contributions/contribution/contribution.component.scss
@@ -1,0 +1,34 @@
+@import "../../../../../styles/variables.scss";
+
+mat-card {
+  border-radius: 0;
+  margin-bottom: 4px;
+}
+
+.publish-status {
+  white-space: nowrap;
+}
+
+.warn {
+  color: $red;
+}
+
+.notice {
+  color: $amber;
+}
+
+.contributor {
+  color: $gray;
+}
+
+.actions {
+  text-align: right;
+  white-space: nowrap;
+  button {
+    text-transform: uppercase;
+    margin-top: 8px;
+    &:last-child {
+      margin-left: 8px;
+    }
+  }
+}
