--- src/app/management/pages/contributions/contributions.component.scss
@@ -1,77 +1,0 @@
-@import "../../../../styles/variables.scss";
-
-table {
-  background-color: white;
-  width: 100%;
-  border-collapse: separate;
-
-  tr {
-    // +1 to accomodate top border on tds (except 1st row which 'uses' border of th)
-    height: 53px;
-    &:first-child {
-      height: 52px;
-    }
-  }
-
-  thead tr {
-    height: 48px;
-    position: sticky;
-    top: 0;
-    background-color: white;
-    z-index: 2;
-  }
-
-  tbody tr {
-    &:hover {
-      background-color: $blue-very-light;
-    }
-  }
-
-  td,
-  th {
-    padding: 8px;
-    text-align: left;
-    font-weight: 400;
-
-    &.icon {
-      text-align: center;
-      width: 40px;
-    }
-  }
-
-  th {
-    color: $gray;
-    vertical-align: bottom;
-    white-space: nowrap;
-    border-bottom: 1px solid $gray-light;
-  }
-
-  td {
-    vertical-align: middle;
-    border-top: 1px solid $gray-light;
-
-    &:last-child {
-      text-align: right;
-    }
-
-    &.actions {
-      button:last-child {
-        margin-left: 8px;
-      }
-    }
-  }
-
-  tbody tr:first-child {
-    td {
-      border: none;
-    }
-  }
-}
-
-.warn {
-  color: $red;
-}
-
-.notice {
-  color: $amber;
-}
