+++ src/app/management/pages/contributions/management-contributions.query.graphql
@@ -22,6 +22,7 @@ query ManagementContributions {
       id
       name
       crag {
+        id
         name
         slug
         publishStatus
@@ -45,6 +46,7 @@ query ManagementContributions {
         publishStatus
       }
     }
+
     id
     publishStatus
   }
